+++ Commands/Web/SetWeb.cs
@@ -5,27 +5,35 @@ using System.Management.Automation;
 namespace SharePointPnP.PowerShell.Commands
 {
     [Cmdlet(VerbsCommon.Set, "PnPWeb")]
-    [CmdletHelp("Sets properties on a web", DetailedDescription = "Sets properties on a web",
-        Category = CmdletHelpCategory.Webs)]
+    [CmdletHelp("Sets properties on a web", 
+        DetailedDescription = "Allows setting various properties on a web",
+        Category = CmdletHelpCategory.Webs,
+        SupportedPlatform = CmdletSupportedPlatform.All)]
     public class SetWeb : PnPWebCmdlet
     {
         [Parameter(Mandatory = false, HelpMessage = "Sets the logo of the web to the current url. If you want to set the logo to a modern team site, use Set-PnPSite -LogoFilePath.")]
         public string SiteLogoUrl;
 
-        [Parameter(Mandatory = false)]
+        [Parameter(Mandatory = false, HelpMessage = "Sets the AlternateCssUrl of the web. Only works for classic pages.")]
         public string AlternateCssUrl;
 
-        [Parameter(Mandatory = false)]
+        [Parameter(Mandatory = false, HelpMessage = "Sets the title of the web")]
         public string Title;
 
-        [Parameter(Mandatory = false)]
+        [Parameter(Mandatory = false, HelpMessage = "Sets the description of the web")]
         public string Description;
 
-        [Parameter(Mandatory = false)]
+        [Parameter(Mandatory = false, HelpMessage = "Sets the MasterUrl of the web. Only works for classic pages.")]
         public string MasterUrl;
 
-        [Parameter(Mandatory = false)]
+        [Parameter(Mandatory = false, HelpMessage = "Sets the CustomMasterUrl of the web. Only works for classic pages.")]
         public string CustomMasterUrl;
+
+        [Parameter(Mandatory = false, HelpMessage = "Defines if the quick launch menu on the left side of modern Team Sites should be shown ($true) or hidden ($false)")]
+        public SwitchParameter QuickLaunchEnabled;
+
+        [Parameter(Mandatory = false, HelpMessage = "Indicates if members of this site can share the site and individual sites with others ($true) or only owners can do this ($false)")]
+        public SwitchParameter MembersCanShare;
 
 #if !ONPREMISES
         [Parameter(Mandatory = false)]
@@ -34,6 +42,17 @@ namespace SharePointPnP.PowerShell.Commands
         [Parameter(Mandatory = false)]
         public SPVariantThemeType HeaderEmphasis = SPVariantThemeType.None;
 
+        [Parameter(Mandatory = false, HelpMessage = "Defines if the navigation menu on a modern site should be enabled for modern audience targeting ($true) or not ($false)")]
+        public SwitchParameter NavAudienceTargetingEnabled;
+
+        [Parameter(Mandatory = false, HelpMessage = "Defines if the navigation menu should be shown as the mega menu ($true) or the smaller sized menu ($false)")]
+        public SwitchParameter MegaMenuEnabled;
+
+        [Parameter(Mandatory = false, HelpMessage = "Defines if Power Automate should be available on lists and document libraries ($false) or if the option should be hidden ($true)")]
+        public SwitchParameter DisablePowerAutomate;
+
+        [Parameter(Mandatory = false, HelpMessage = "Defines if comments on modern site pages should be enabled by default ($false) or they should be hidden ($true)")]
+        public SwitchParameter CommentsOnSitePagesDisabled;
 #endif
 
         protected override void ExecuteCmdlet()
@@ -73,6 +92,16 @@ namespace SharePointPnP.PowerShell.Commands
                         SelectedWeb.CustomMasterUrl = CustomMasterUrl;
                         dirty = true;
                         break;
+
+                    case nameof(QuickLaunchEnabled):
+                        SelectedWeb.QuickLaunchEnabled = QuickLaunchEnabled.ToBool();
+                        dirty = true;
+                        break;
+
+                    case nameof(MembersCanShare):
+                        SelectedWeb.MembersCanShare = MembersCanShare.ToBool();
+                        dirty = true;
+                        break;
 #if !ONPREMISES
                     case nameof(HeaderLayout):
                         SelectedWeb.HeaderLayout = HeaderLayout;
@@ -81,6 +110,26 @@ namespace SharePointPnP.PowerShell.Commands
 
                     case nameof(HeaderEmphasis):
                         SelectedWeb.HeaderEmphasis = HeaderEmphasis;
+                        dirty = true;
+                        break;
+
+                    case nameof(NavAudienceTargetingEnabled):
+                        SelectedWeb.NavAudienceTargetingEnabled = NavAudienceTargetingEnabled.ToBool();
+                        dirty = true;
+                        break;
+
+                    case nameof(MegaMenuEnabled):
+                        SelectedWeb.MegaMenuEnabled = MegaMenuEnabled.ToBool();
+                        dirty = true;
+                        break;
+
+                    case nameof(DisablePowerAutomate):
+                        SelectedWeb.DisableFlows = DisablePowerAutomate.ToBool();
+                        dirty = true;
+                        break;
+
+                    case nameof(CommentsOnSitePagesDisabled):
+                        SelectedWeb.CommentsOnSitePagesDisabled = CommentsOnSitePagesDisabled.ToBool();
                         dirty = true;
                         break;
 #endif
@@ -94,5 +143,4 @@ namespace SharePointPnP.PowerShell.Commands
             }
         }
     }
-
 }